+++ crates/node/src/command.rs
@@ -1,14 +1,15 @@
-use std::path::PathBuf;
-
 use frame_benchmarking_cli::{BenchmarkCmd, ExtrinsicFactory, SUBSTRATE_REFERENCE_HARDWARE};
 use madara_runtime::Block;
-use mc_data_availability::DaLayer;
-use pallet_starknet::utils;
-use sc_cli::{ChainSpec, RpcMethods, RuntimeVersion, SubstrateCli};
+use sc_cli::{ChainSpec, RuntimeVersion, SubstrateCli};
 
 use crate::benchmarking::{inherent_benchmark_data, RemarkBuilder};
-use crate::cli::{Cli, Subcommand, Testnet};
-use crate::{chain_spec, constants, service};
+use crate::cli::{Cli, Subcommand};
+use crate::commands::run_node;
+use crate::constants::DEV_CHAIN_ID;
+#[cfg(feature = "sharingan")]
+use crate::constants::SHARINGAN_CHAIN_ID;
+use crate::{chain_spec, service};
+
 impl SubstrateCli for Cli {
     fn impl_name() -> String {
         "Madara Node".into()
@@ -36,17 +37,20 @@ impl SubstrateCli for Cli {
 
     fn load_spec(&self, id: &str) -> Result<Box<dyn sc_service::ChainSpec>, String> {
         Ok(match id {
-            "dev" => {
-                let enable_manual_seal = self.sealing.map(|_| true);
-                Box::new(chain_spec::development_config(
-                    enable_manual_seal,
-                    self.run.madara_path.clone().expect("Failed retrieving madara_path"),
-                )?)
+            DEV_CHAIN_ID => {
+                let enable_manual_seal = self.run.sealing.map(|_| true);
+                let base_path = self.run.base_path().map_err(|e| e.to_string())?;
+                Box::new(chain_spec::development_config(enable_manual_seal, base_path)?)
             }
-            "" | "local" | "madara-local" => Box::new(chain_spec::local_testnet_config(
-                self.run.madara_path.clone().expect("Failed retrieving madara_path"),
+            #[cfg(feature = "sharingan")]
+            SHARINGAN_CHAIN_ID => Box::new(chain_spec::ChainSpec::from_json_bytes(
+                &include_bytes!("../../../configs/chain-specs/testnet-sharingan-raw.json")[..],
             )?),
-            path => Box::new(chain_spec::ChainSpec::from_json_file(std::path::PathBuf::from(path))?),
+            "" | "local" | "madara-local" => {
+                let base_path = self.run.base_path().map_err(|e| e.to_string())?;
+                Box::new(chain_spec::local_testnet_config(base_path, id)?)
+            }
+            path_or_url => Box::new(chain_spec::ChainSpec::from_json_file(std::path::PathBuf::from(path_or_url))?),
         })
     }
 
@@ -57,108 +61,47 @@ impl SubstrateCli for Cli {
 
 /// Parse and run command line arguments
 pub fn run() -> sc_cli::Result<()> {
-    let mut cli = Cli::from_args();
+    let cli = Cli::from_args();
 
-    // alias madara_path <> base_path
-    // TODO also alias tmp (tmp generates random base_paths that are not specified within
-    // the command)
-    let madara_path = match (cli.run.madara_path.clone(), cli.run.run_cmd.shared_params.base_path.clone()) {
-        (Some(madara_path), _) => {
-            cli.run.run_cmd.shared_params.base_path = Some(madara_path.clone());
-            madara_path.to_str().unwrap().to_string()
-        }
-        (_, Some(base_path)) => {
-            cli.run.madara_path = Some(base_path.clone());
-            base_path.to_str().unwrap().to_string()
-        }
-        _ => {
-            let home_path = std::env::var("HOME").unwrap_or(std::env::var("USERPROFILE").unwrap_or(".".into()));
-            let path = format!("{}/.madara", home_path);
-            cli.run.run_cmd.shared_params.base_path = Some((path.clone()).into());
-            cli.run.madara_path = Some((path.clone()).into());
-            path
-        }
-    };
-
-    if let Some(genesis_url) = cli.run.genesis_url.clone() {
-        // can't copy extra genesis-assets atm
-        // we can reuse #982 to create the standard to fetch relevant files
-        utils::fetch_from_url(genesis_url, madara_path.clone() + "/configs/genesis-assets")?;
-    } else {
-        // TODO confirm with the CI that we are fetching all and fetch dynamically
-        // Issue #982
-        for file in constants::GENESIS_ASSETS_FILES {
-            let src_path = utils::get_project_path();
-            if let Ok(src_path) = src_path {
-                let src_path = src_path + "/configs/genesis-assets/" + file;
-                utils::copy_from_filesystem(src_path, madara_path.clone() + "/genesis-assets")?;
-            } else {
-                utils::fetch_from_url(
-                    constants::GENESIS_ASSETS_URL.to_string() + file,
-                    madara_path.clone() + "/genesis-assets",
-                )?;
-            }
-        }
-    }
-
-    // TODO confirm with the CI that we are fetching all and fetch dynamically
-    // Issue #982
-    for file in constants::CAIRO_CONTRACTS_FILES {
-        let src_path = utils::get_project_path();
-        if let Ok(src_path) = src_path {
-            let src_path = src_path + "/configs/cairo-contracts/" + file;
-            utils::copy_from_filesystem(src_path, madara_path.clone() + "/cairo-contracts")?;
-        } else {
-            utils::fetch_from_url(
-                constants::CAIRO_CONTRACTS_URL.to_string() + file,
-                madara_path.clone() + "/cairo-contracts",
-            )?;
-        }
-    }
-
-    if let (Some(chain_spec_url), None) = (cli.run.chain_spec_url.clone(), cli.run.testnet) {
-        utils::fetch_from_url(chain_spec_url, madara_path.clone() + "/chain-specs")?;
-    }
-
-    match &cli.subcommand {
-        Some(Subcommand::Key(cmd)) => cmd.run(&cli),
-        Some(Subcommand::BuildSpec(cmd)) => {
+    match cli.subcommand {
+        Some(Subcommand::Key(ref cmd)) => cmd.run(&cli),
+        Some(Subcommand::BuildSpec(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.sync_run(|config| cmd.run(config.chain_spec, config.network))
         }
-        Some(Subcommand::CheckBlock(cmd)) => {
+        Some(Subcommand::CheckBlock(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, import_queue, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, import_queue), task_manager))
             })
         }
-        Some(Subcommand::ExportBlocks(cmd)) => {
+        Some(Subcommand::ExportBlocks(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, _, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, config.database), task_manager))
             })
         }
-        Some(Subcommand::ExportState(cmd)) => {
+        Some(Subcommand::ExportState(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, _, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, config.chain_spec), task_manager))
             })
         }
-        Some(Subcommand::ImportBlocks(cmd)) => {
+        Some(Subcommand::ImportBlocks(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, _, import_queue, task_manager, _) = service::new_chain_ops(&mut config)?;
                 Ok((cmd.run(client, import_queue), task_manager))
             })
         }
-        Some(Subcommand::PurgeChain(cmd)) => {
+        Some(Subcommand::PurgeChain(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.sync_run(|config| cmd.run(config.database))
         }
-        Some(Subcommand::Revert(cmd)) => {
+        Some(Subcommand::Revert(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.async_run(|mut config| {
                 let (client, backend, _, task_manager, _) = service::new_chain_ops(&mut config)?;
@@ -169,7 +112,7 @@ pub fn run() -> sc_cli::Result<()> {
                 Ok((cmd.run(client, backend, Some(aux_revert)), task_manager))
             })
         }
-        Some(Subcommand::Benchmark(cmd)) => {
+        Some(Subcommand::Benchmark(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
 
             runner.sync_run(|mut config| {
@@ -234,68 +177,11 @@ pub fn run() -> sc_cli::Result<()> {
         Some(Subcommand::TryRuntime) => Err("TryRuntime wasn't enabled when building the node. You can enable it \
                                              with `--features try-runtime`."
             .into()),
-        Some(Subcommand::ChainInfo(cmd)) => {
+        Some(Subcommand::ChainInfo(ref cmd)) => {
             let runner = cli.create_runner(cmd)?;
             runner.sync_run(|config| cmd.run::<Block>(&config))
         }
-        None => {
-            // create a reproducible dev environment
-            if cli.run.run_cmd.shared_params.dev {
-                cli.run.run_cmd.shared_params.dev = false;
-                cli.run.run_cmd.shared_params.chain = Some("dev".to_string());
-
-                cli.run.run_cmd.force_authoring = true;
-                cli.run.run_cmd.alice = true;
-
-                // we can't set `--rpc-cors=all`, so it needs to be set manually if we want to connect with external
-                // hosts
-                cli.run.run_cmd.rpc_external = true;
-                cli.run.run_cmd.rpc_methods = RpcMethods::Unsafe;
-            }
-
-            cli.run.run_cmd.network_params.node_key_params.node_key_file =
-                Some((madara_path.clone() + "/p2p-key.ed25519").into());
-
-            if let Some(Testnet::Sharingan) = cli.run.testnet {
-                let src_path = utils::get_project_path();
-                if let Ok(src_path) = src_path {
-                    let src_path = src_path + "/configs/chain-specs/testnet-sharingan-raw.json";
-                    utils::copy_from_filesystem(src_path, madara_path.clone() + "/chain-specs")?;
-                } else {
-                    utils::fetch_from_url(
-                        constants::SHARINGAN_CHAIN_SPEC_URL.to_string(),
-                        madara_path.clone() + "/chain-specs",
-                    )?;
-                }
-
-                cli.run.run_cmd.shared_params.chain =
-                    Some(madara_path.clone() + "/chain-specs/testnet-sharingan-raw.json");
-
-                // This should go apply to all testnets when applying a match pattern
-                cli.run.run_cmd.rpc_external = true;
-                cli.run.run_cmd.rpc_methods = RpcMethods::Unsafe;
-            }
-
-            let da_config: Option<(DaLayer, PathBuf)> = match cli.run.da_layer {
-                Some(da_layer) => {
-                    let da_path = std::path::PathBuf::from(madara_path.clone() + "/da-config.json");
-                    if !da_path.exists() {
-                        log::info!("{} does not contain DA config", madara_path);
-                        return Err("DA config not available".into());
-                    }
-
-                    Some((da_layer, da_path))
-                }
-                None => {
-                    log::info!("madara initialized w/o da layer");
-                    None
-                }
-            };
-
-            let runner = cli.create_runner(&cli.run.run_cmd)?;
-            runner.run_node_until_exit(|config| async move {
-                service::new_full(config, cli.sealing, da_config).map_err(sc_cli::Error::Service)
-            })
-        }
+        Some(Subcommand::Setup(ref cmd)) => cmd.run(),
+        None => run_node(cli),
     }
 }